14 abril 2020

+++ CaribeanJudge/a+b.cpp
@@ -4,8 +4,7 @@ using namespace std;
 
 int main()
 {
-
-	int a, b;
+	long long a, b;
 	cin >> a >> b;
 	cout << a + b << '\n';
 	return 0;
